avoid needless Node module externalization + split plugin utils
