+++ src/plugins/namespace-map.ts
@@ -1,0 +1,17 @@
+/**
+ * Maps namespaces that deviate from the file-name
+ *
+ * @remarks expects file-name as value and custom-namespace as key
+ */
+export const namespaceMap = {
+  titles: "trainer-titles",
+  moveTriggers: "move-trigger",
+  abilityTriggers: "ability-trigger",
+  battlePokemonForm: "pokemon-form-battle",
+  miscDialogue: "dialogue-misc",
+  battleSpecDialogue: "dialogue-final-boss",
+  doubleBattleDialogue: "dialogue-double-battle",
+  splashMessages: "splash-texts",
+  mysteryEncounterMessages: "mystery-encounter-texts",
+  biome: "biomes",
+};
