+++ vite.config.ts
@@ -65,10 +65,10 @@ export const sharedConfig: UserConfigFnPromise = async ({ mode }) => {
 
   if (!process.env.MERGE_REPORTS) {
     opts.plugins = [
+      tsconfigPaths(),
       (await import("./src/plugins/vite/vite-minify-json-plugin")).minifyPublicJsonFiles(),
       (await import("./src/plugins/vite/namespaces-i18n-plugin")).LocaleNamespace(),
       (await import("unplugin-inline-enum/vite")).default({ scanDir: "src" }),
-      tsconfigPaths(),
     ];
   }
   return opts;
